requirements.txt
# =============================================================================
# Form Oluşturma Uygulaması - Bağımlılıklar
# =============================================================================
# Kurulum: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Temel UI
# -----------------------------------------------------------------------------
PyQt5>=5.15.0

# -----------------------------------------------------------------------------
# Belge İşleme
# -----------------------------------------------------------------------------
python-docx>=0.8.11
docxcompose>=1.3.0

# -----------------------------------------------------------------------------
# Geliştirme Araçları (opsiyonel)
# -----------------------------------------------------------------------------
# pytest>=7.0.0          # Test framework
# pytest-qt>=4.0.0       # PyQt5 test desteği
# black>=22.0.0          # Kod formatlama
# flake8>=4.0.0          # Kod kalite kontrolü
# mypy>=0.950            # Tip kontrolü

# -----------------------------------------------------------------------------
# Gelecek Özellikler (henüz gerekli değil)
# -----------------------------------------------------------------------------
# google-api-python-client>=2.0.0    # Google Drive entegrasyonu
# google-auth-httplib2>=0.1.0
# google-auth-oauthlib>=0.5.0
